Divide generic generation in submodules
