+++ mesh_to_sdf/src/generate/generic/mod.rs
@@ -1,0 +1,6 @@
+//! Module containing the different `generate_sdf` functions, one for each acceleration structure.
+
+pub(crate) mod bvh;
+pub(crate) mod default;
+pub(crate) mod rtree;
+pub(crate) mod rtree_bvh;
