+++ mesh_to_sdf/src/lib.rs
@@ -122,16 +122,15 @@
 //!
 //! #### Benchmarks
 //!
-//! [`generate_grid_sdf`] is much faster than [`generate_sdf`] and should be used whenever possible.
-//! [`generate_sdf`] does not allocate memory (except for the result array) but is slow. A faster implementation is planned for the future.
-//!
 //! [`SignMethod::Raycast`] is slightly slower than [`SignMethod::Normal`] but is robust and should be used whenever possible (~1% in [`generate_grid_sdf`], ~10% in [`generate_sdf`]).
-use std::{boxed::Box, cmp::Ordering};
+use std::boxed::Box;
 
-use bvh::{bounding_hierarchy::BoundingHierarchy, bvh::Bvh};
-use bvh_ext::BvhDistance;
 use itertools::Itertools;
-use rayon::prelude::*;
+
+use generate::generic::{
+    bvh::generate_sdf_bvh, default::generate_sdf_default, rtree::generate_sdf_rtree,
+    rtree_bvh::generate_sdf_rtree_bvh,
+};
 
 mod bvh_ext;
 mod generate;
@@ -260,42 +259,6 @@ fn compare_distances(a: f32, b: f32) -> std::cmp::Ordering {
     }
 }
 
-#[derive(Clone)]
-// TODO: split in BvhNode / RtreeNode / RtreeBvhNode to avoid allocating for nothing.
-// This is because bvh only needs indices but rtree needs vertices.
-struct BvhNode<V: Point> {
-    vertex_indices: (usize, usize, usize),
-    vertices: (V, V, V),
-    node_index: usize,
-    bounding_box: (V, V),
-}
-
-impl<V: Point> bvh::aabb::Bounded<f32, 3> for BvhNode<V> {
-    fn aabb(&self) -> bvh::aabb::Aabb<f32, 3> {
-        let min = nalgebra::Point3::new(
-            self.bounding_box.0.x(),
-            self.bounding_box.0.y(),
-            self.bounding_box.0.z(),
-        );
-        let max = nalgebra::Point3::new(
-            self.bounding_box.1.x(),
-            self.bounding_box.1.y(),
-            self.bounding_box.1.z(),
-        );
-        bvh::aabb::Aabb::with_bounds(min, max)
-    }
-}
-
-impl<V: Point> bvh::bounding_hierarchy::BHShape<f32, 3> for BvhNode<V> {
-    fn set_bh_node_index(&mut self, index: usize) {
-        self.node_index = index;
-    }
-
-    fn bh_node_index(&self) -> usize {
-        self.node_index
-    }
-}
-
 /// Generate a signed distance field from a mesh.
 /// Query points are expected to be in the same space as the mesh.
 ///
@@ -345,812 +308,5 @@ where
         }
         AccelerationMethod::Rtree => generate_sdf_rtree(vertices, indices, query_points),
         AccelerationMethod::RtreeBvh => generate_sdf_rtree_bvh(vertices, indices, query_points),
-    }
-}
-
-/// Generate a signed distance field from a mesh using a bvh.
-/// Query points are expected to be in the same space as the mesh.
-///
-/// Returns a vector of signed distances.
-/// Queries outside the mesh will have a positive distance, and queries inside the mesh will have a negative distance.
-fn generate_sdf_bvh<V, I>(
-    vertices: &[V],
-    indices: Topology<I>,
-    query_points: &[V],
-    sign_method: SignMethod,
-) -> Vec<f32>
-where
-    V: Point,
-    I: Copy + Into<u32> + Sync + Send,
-{
-    let mut bvh_nodes = Topology::get_triangles(vertices, indices)
-        .map(|triangle| BvhNode {
-            vertex_indices: triangle,
-            vertices: (
-                vertices[triangle.0],
-                vertices[triangle.1],
-                vertices[triangle.2],
-            ),
-            node_index: 0,
-            bounding_box: geo::triangle_bounding_box(
-                &vertices[triangle.0],
-                &vertices[triangle.1],
-                &vertices[triangle.2],
-            ),
-        })
-        .collect_vec();
-
-    let bvh = Bvh::build_par(&mut bvh_nodes);
-
-    query_points
-        .par_iter()
-        .map(|point| {
-            let bvh_indices = bvh.nearest_candidates(point, &bvh_nodes);
-
-            let mut min_dist = f32::MAX;
-            if sign_method == SignMethod::Normal {
-                for index in &bvh_indices {
-                    let triangle = &bvh_nodes[*index];
-                    let a = &vertices[triangle.vertex_indices.0];
-                    let b = &vertices[triangle.vertex_indices.1];
-                    let c = &vertices[triangle.vertex_indices.2];
-                    let distance = geo::point_triangle_signed_distance(point, a, b, c);
-
-                    if compare_distances(min_dist, distance) == Ordering::Greater {
-                        min_dist = distance;
-                    }
-                }
-                min_dist
-            } else {
-                for index in &bvh_indices {
-                    let triangle = &bvh_nodes[*index];
-                    let a = &vertices[triangle.vertex_indices.0];
-                    let b = &vertices[triangle.vertex_indices.1];
-                    let c = &vertices[triangle.vertex_indices.2];
-                    let distance = geo::point_triangle_distance(point, a, b, c);
-
-                    min_dist = min_dist.min(distance);
-                }
-
-                let alignments = [
-                    (geo::GridAlign::X, nalgebra::Vector3::new(1.0, 0.0, 0.0)),
-                    (geo::GridAlign::Y, nalgebra::Vector3::new(0.0, 1.0, 0.0)),
-                    (geo::GridAlign::Z, nalgebra::Vector3::new(0.0, 0.0, 1.0)),
-                ];
-
-                let mut insides = 0;
-                for (alignment, direction) in alignments {
-                    let ray = bvh::ray::Ray::new(
-                        nalgebra::Point3::new(point.x(), point.y(), point.z()),
-                        direction,
-                    );
-                    let mut intersection_count = 0;
-                    let hitcast = bvh.traverse(&ray, &bvh_nodes);
-                    for bvh_node in hitcast {
-                        let a = &vertices[bvh_node.vertex_indices.0];
-                        let b = &vertices[bvh_node.vertex_indices.1];
-                        let c = &vertices[bvh_node.vertex_indices.2];
-                        let intersect =
-                            geo::ray_triangle_intersection_aligned(point, [a, b, c], alignment);
-                        if intersect.is_some() {
-                            intersection_count += 1;
-                        }
-                    }
-
-                    if intersection_count % 2 == 1 {
-                        insides += 1;
-                    }
-                }
-
-                // Return inside if at least two are insides.
-                if insides > 1 {
-                    -min_dist
-                } else {
-                    min_dist
-                }
-            }
-        })
-        .collect()
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Hash)]
-struct PointWrapper<V>(V);
-
-impl<V> rstar::Point for PointWrapper<V>
-where
-    V: Point,
-{
-    type Scalar = f32;
-
-    const DIMENSIONS: usize = 3;
-
-    fn generate(mut generator: impl FnMut(usize) -> Self::Scalar) -> Self {
-        let x = generator(0);
-        let y = generator(1);
-        let z = generator(2);
-        PointWrapper(V::new(x, y, z))
-    }
-
-    fn nth(&self, index: usize) -> Self::Scalar {
-        match index {
-            0 => self.0.x(),
-            1 => self.0.y(),
-            2 => self.0.z(),
-            _ => panic!("Index out of bounds"),
-        }
-    }
-
-    fn nth_mut(&mut self, index: usize) -> &mut Self::Scalar {
-        match index {
-            0 => self.0.x_mut(),
-            1 => self.0.y_mut(),
-            2 => self.0.z_mut(),
-            _ => panic!("Index out of bounds"),
-        }
-    }
-}
-
-impl<V: Point> rstar::RTreeObject for BvhNode<V> {
-    type Envelope = rstar::AABB<PointWrapper<V>>;
-
-    fn envelope(&self) -> Self::Envelope {
-        rstar::AABB::from_corners(
-            PointWrapper(self.bounding_box.0),
-            PointWrapper(self.bounding_box.1),
-        )
-    }
-}
-
-impl<V: Point> rstar::PointDistance for BvhNode<V> {
-    // Required method
-    fn distance_2(
-        &self,
-        point: &<Self::Envelope as rstar::Envelope>::Point,
-    ) -> <<Self::Envelope as rstar::Envelope>::Point as rstar::Point>::Scalar {
-        geo::point_triangle_distance2(
-            &point.0,
-            &self.vertices.0,
-            &self.vertices.1,
-            &self.vertices.2,
-        )
-    }
-}
-
-/// Generate a signed distance field from a mesh using an r-tree.
-/// Query points are expected to be in the same space as the mesh.
-///
-/// Returns a vector of signed distances.
-/// Queries outside the mesh will have a positive distance, and queries inside the mesh will have a negative distance.
-///
-/// This method is only compatible with normal sign method.
-/// If you want to use raycasting, use `AccelerationMethod::RtreeBvh` instead.
-fn generate_sdf_rtree<V, I>(vertices: &[V], indices: Topology<I>, query_points: &[V]) -> Vec<f32>
-where
-    V: Point,
-    I: Copy + Into<u32> + Sync + Send,
-{
-    let bvh_nodes = Topology::get_triangles(vertices, indices)
-        .map(|triangle| BvhNode {
-            vertex_indices: triangle,
-            vertices: (
-                vertices[triangle.0],
-                vertices[triangle.1],
-                vertices[triangle.2],
-            ),
-            node_index: 0,
-            bounding_box: geo::triangle_bounding_box(
-                &vertices[triangle.0],
-                &vertices[triangle.1],
-                &vertices[triangle.2],
-            ),
-        })
-        .collect_vec();
-
-    let rtree = rstar::RTree::bulk_load(bvh_nodes);
-
-    query_points
-        .par_iter()
-        .map(|point| {
-            let nearest = rtree.nearest_neighbor(&PointWrapper(*point));
-            let nearest = nearest.unwrap();
-            geo::point_triangle_signed_distance(
-                point,
-                &nearest.vertices.0,
-                &nearest.vertices.1,
-                &nearest.vertices.2,
-            )
-        })
-        .collect()
-}
-
-/// Generate a signed distance field from a mesh using an r-tree for nearest neighbor search and a bvh for ray intersection.
-/// Query points are expected to be in the same space as the mesh.
-///
-/// Returns a vector of signed distances.
-/// Queries outside the mesh will have a positive distance, and queries inside the mesh will have a negative distance.
-fn generate_sdf_rtree_bvh<V, I>(
-    vertices: &[V],
-    indices: Topology<I>,
-    query_points: &[V],
-) -> Vec<f32>
-where
-    V: Point,
-    I: Copy + Into<u32> + Sync + Send,
-{
-    let mut bvh_nodes = Topology::get_triangles(vertices, indices)
-        .map(|triangle| BvhNode {
-            vertex_indices: triangle,
-            vertices: (
-                vertices[triangle.0],
-                vertices[triangle.1],
-                vertices[triangle.2],
-            ),
-            node_index: 0,
-            bounding_box: geo::triangle_bounding_box(
-                &vertices[triangle.0],
-                &vertices[triangle.1],
-                &vertices[triangle.2],
-            ),
-        })
-        .collect_vec();
-
-    if bvh_nodes.is_empty() {
-        return vec![];
-    }
-
-    let rtree = rstar::RTree::bulk_load(bvh_nodes.clone());
-    let bvh = Bvh::build_par(&mut bvh_nodes);
-
-    query_points
-        .par_iter()
-        .map(|point| {
-            let nearest = rtree.nearest_neighbor(&PointWrapper(*point));
-            let nearest = nearest.unwrap(); // tree isn't empty.
-
-            let dist = geo::point_triangle_distance(
-                point,
-                &nearest.vertices.0,
-                &nearest.vertices.1,
-                &nearest.vertices.2,
-            );
-
-            let alignments = [
-                (geo::GridAlign::X, nalgebra::Vector3::new(1.0, 0.0, 0.0)),
-                (geo::GridAlign::Y, nalgebra::Vector3::new(0.0, 1.0, 0.0)),
-                (geo::GridAlign::Z, nalgebra::Vector3::new(0.0, 0.0, 1.0)),
-            ];
-
-            let mut insides = 0;
-            for (alignment, direction) in alignments {
-                let ray = bvh::ray::Ray::new(
-                    nalgebra::Point3::new(point.x(), point.y(), point.z()),
-                    direction,
-                );
-                let mut intersection_count = 0;
-                let hitcast = bvh.traverse(&ray, &bvh_nodes);
-                for bvh_node in hitcast {
-                    let a = &bvh_node.vertices.0;
-                    let b = &bvh_node.vertices.1;
-                    let c = &bvh_node.vertices.2;
-                    let intersect =
-                        geo::ray_triangle_intersection_aligned(point, [a, b, c], alignment);
-                    if intersect.is_some() {
-                        intersection_count += 1;
-                    }
-                }
-
-                if intersection_count % 2 == 1 {
-                    insides += 1;
-                }
-            }
-
-            // Return inside if at least two are insides.
-            if insides > 1 {
-                -dist
-            } else {
-                dist
-            }
-        })
-        .collect()
-}
-
-/// Generate a signed distance field from a mesh.
-/// Query points are expected to be in the same space as the mesh.
-///
-/// Returns a vector of signed distances.
-/// Queries outside the mesh will have a positive distance, and queries inside the mesh will have a negative distance.
-fn generate_sdf_default<V, I>(
-    vertices: &[V],
-    indices: Topology<I>,
-    query_points: &[V],
-    sign_method: SignMethod,
-) -> Vec<f32>
-where
-    V: Point,
-    I: Copy + Into<u32> + Sync + Send,
-{
-    // For each query point, we compute the distance to each triangle.
-    // sign is estimated by comparing the normal to the direction.
-    // when two triangles give the same distance (wrt floating point errors),
-    // we keep the one with positive distance since to be inside means to be inside all triangles.
-    // whereas to be outside means to be outside at least one triangle.
-    // see `compare_distances` for more details.
-    query_points
-        .par_iter()
-        .map(|query| {
-            Topology::get_triangles(vertices, indices)
-                .map(|(i, j, k)| (&vertices[i], &vertices[j], &vertices[k]))
-                .map(|(a, b, c)| match sign_method {
-                    // Raycast: returns (distance, ray_intersection)
-                    SignMethod::Raycast => (
-                        geo::point_triangle_distance(query, a, b, c),
-                        geo::ray_triangle_intersection_aligned(query, [a, b, c], geo::GridAlign::X)
-                            .is_some(),
-                    ),
-                    // Normal: returns (signed_distance, false)
-                    SignMethod::Normal => {
-                        (geo::point_triangle_signed_distance(query, a, b, c), false)
-                    }
-                })
-                .fold(
-                    (f32::MAX, 0),
-                    |(min_distance, intersection_count), (distance, ray_intersection)| {
-                        match sign_method {
-                            SignMethod::Raycast => (
-                                min_distance.min(distance),
-                                intersection_count + ray_intersection as u32,
-                            ),
-                            SignMethod::Normal => (
-                                match compare_distances(distance, min_distance) {
-                                    std::cmp::Ordering::Less => distance,
-                                    _ => min_distance,
-                                },
-                                intersection_count,
-                            ),
-                        }
-                    },
-                )
-        })
-        .map(|(distance, intersection_count)| {
-            if intersection_count % 2 == 0 {
-                distance
-            } else {
-                // can only be odd if in raycast mode
-                -distance
-            }
-        })
-        .collect()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_generate() {
-        let model = &easy_gltf::load("assets/suzanne.glb").unwrap()[0].models[0];
-        let vertices = model.vertices().iter().map(|v| v.position).collect_vec();
-        let indices = model.indices().unwrap();
-        let query_points = [
-            cgmath::Vector3::new(0.0, 0.0, 0.0),
-            cgmath::Vector3::new(1.0, 1.0, 1.0),
-            cgmath::Vector3::new(0.1, 0.2, 0.2),
-        ];
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::None(SignMethod::Normal),
-        );
-
-        // pysdf [0.45216727 -0.6997909   0.45411023] # negative is outside in pysdf
-        // mesh_to_sdf [-0.40961263  0.6929414  -0.46345082] # negative is inside in mesh_to_sdf
-        let baseline = [-0.42, 0.69, -0.46];
-
-        // make sure the results are close enough.
-        // the results are not exactly the same because the algorithm is not the same and baselines might not be exact.
-        // this is mostly to make sure the results are not completely off.
-        for (sdf, baseline) in sdf.iter().zip(baseline.iter()) {
-            assert!((sdf - baseline).abs() < 0.1);
-        }
-    }
-
-    #[test]
-    fn test_generate_grid() {
-        // make sure generate_grid_sdf returns the same result as generate_sdf.
-        // assumes generate_sdf is properly tested and correct.
-        let vertices: Vec<[f32; 3]> = vec![[0., 1., 0.], [1., 2., 3.], [1., 3., 4.], [2., 0., 0.]];
-        let indices: Vec<u32> = vec![0, 1, 2, 1, 2, 3];
-        let grid = Grid::from_bounding_box(&[0., 0., 0.], &[5., 5., 5.], [5, 5, 5]);
-        let mut query_points = Vec::new();
-        for x in 0..grid.get_cell_count()[0] {
-            for y in 0..grid.get_cell_count()[1] {
-                for z in 0..grid.get_cell_count()[2] {
-                    query_points.push(grid.get_cell_center(&[x, y, z]));
-                }
-            }
-        }
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(&indices)),
-            &query_points,
-            AccelerationMethod::None(SignMethod::Raycast),
-        );
-        let grid_sdf = generate_grid_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(&indices)),
-            &grid,
-            SignMethod::Raycast,
-        );
-
-        // Test against generate_sdf
-        for (i, (sdf, grid_sdf)) in sdf.iter().zip(grid_sdf.iter()).enumerate() {
-            assert_eq!(sdf, grid_sdf, "i: {}", i);
-        }
-    }
-
-    #[test]
-    fn test_generate_bvh() {
-        let model = &easy_gltf::load("assets/suzanne.glb").unwrap()[0].models[0];
-        let vertices = model.vertices().iter().map(|v| v.position).collect_vec();
-        let indices = model.indices().unwrap();
-        let query_points = [
-            cgmath::Vector3::new(0.01, 0.01, 0.5),
-            cgmath::Vector3::new(1.0, 1.0, 1.0),
-            cgmath::Vector3::new(0.1, 0.2, 0.2),
-            cgmath::Vector3::new(1.1, 2.2, 5.2),
-            cgmath::Vector3::new(-0.1, 0.2, -0.2),
-        ];
-
-        let bvh_sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::Bvh(SignMethod::Raycast),
-        );
-
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::None(SignMethod::Raycast),
-        );
-
-        for (idx, (bvh, sdf)) in bvh_sdf.iter().zip(sdf.iter()).enumerate() {
-            assert!(
-                (bvh - sdf).abs() < 0.01,
-                "{:?}: {} != {}",
-                query_points[idx],
-                bvh,
-                sdf
-            );
-        }
-    }
-
-    #[test]
-    fn test_generate_bvh_big() {
-        let model = &easy_gltf::load("assets/suzanne.glb").unwrap()[0].models[0];
-        let vertices = model.vertices().iter().map(|v| v.position).collect_vec();
-        let indices = model.indices().unwrap();
-
-        let bbox_min = vertices.iter().fold(
-            cgmath::Vector3::new(f32::MAX, f32::MAX, f32::MAX),
-            |acc, v| cgmath::Vector3 {
-                x: acc.x.min(v.x),
-                y: acc.y.min(v.y),
-                z: acc.z.min(v.z),
-            },
-        );
-        let bbox_max = vertices.iter().fold(
-            cgmath::Vector3::new(-f32::MAX, -f32::MAX, -f32::MAX),
-            |acc, v| cgmath::Vector3 {
-                x: acc.x.max(v.x),
-                y: acc.y.max(v.y),
-                z: acc.z.max(v.z),
-            },
-        );
-
-        let grid = Grid::from_bounding_box(&bbox_min, &bbox_max, [32, 32, 32]);
-        let mut query_points = Vec::new();
-        for x in 0..grid.get_cell_count()[0] {
-            for y in 0..grid.get_cell_count()[1] {
-                for z in 0..grid.get_cell_count()[2] {
-                    query_points.push(grid.get_cell_center(&[x, y, z]));
-                }
-            }
-        }
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::Bvh(SignMethod::Raycast),
-        );
-        let grid_sdf = generate_grid_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &grid,
-            SignMethod::Raycast,
-        );
-
-        // Test against generate_sdf
-        // TODO: sometimes fails
-        // thread 'tests::test_generate_bvh_big' panicked at mesh_to_sdf\src\lib.rs:1232:13:
-        // i: 17435: 0.0076956493 0.030284861
-        for (i, (sdf, grid_sdf)) in sdf.iter().zip(grid_sdf.iter()).enumerate() {
-            assert!(
-                (sdf - grid_sdf).abs() < 0.01,
-                "cell: {:?}: {} {}",
-                grid.get_cell_integer_coordinates(i),
-                sdf,
-                grid_sdf
-            );
-        }
-    }
-
-    #[test]
-    fn test_generate_bvh_normal() {
-        let model = &easy_gltf::load("assets/suzanne.glb").unwrap()[0].models[0];
-        let vertices = model.vertices().iter().map(|v| v.position).collect_vec();
-        let indices = model.indices().unwrap();
-
-        let bbox_min = vertices.iter().fold(
-            cgmath::Vector3::new(f32::MAX, f32::MAX, f32::MAX),
-            |acc, v| cgmath::Vector3 {
-                x: acc.x.min(v.x),
-                y: acc.y.min(v.y),
-                z: acc.z.min(v.z),
-            },
-        );
-        let bbox_max = vertices.iter().fold(
-            cgmath::Vector3::new(-f32::MAX, -f32::MAX, -f32::MAX),
-            |acc, v| cgmath::Vector3 {
-                x: acc.x.max(v.x),
-                y: acc.y.max(v.y),
-                z: acc.z.max(v.z),
-            },
-        );
-
-        let grid = Grid::from_bounding_box(&bbox_min, &bbox_max, [16, 16, 16]);
-        let mut query_points = Vec::new();
-        for x in 0..grid.get_cell_count()[0] {
-            for y in 0..grid.get_cell_count()[1] {
-                for z in 0..grid.get_cell_count()[2] {
-                    query_points.push(grid.get_cell_center(&[x, y, z]));
-                }
-            }
-        }
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::Bvh(SignMethod::Normal),
-        );
-        let grid_sdf = generate_grid_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &grid,
-            SignMethod::Normal,
-        );
-
-        // Test against generate_sdf
-        for (i, (sdf, grid_sdf)) in sdf.iter().zip(grid_sdf.iter()).enumerate() {
-            // TODO: sometimes fails
-            // thread 'tests::test_generate_bvh_big' panicked at mesh_to_sdf\src\lib.rs:1232:13:
-            // i: 17435: 0.0076956493 0.030284861
-            // i: 1742: 0.09342232 -0.094851956
-            assert!(
-                (sdf - grid_sdf).abs() < 0.01,
-                "i: {}: {} {}",
-                i,
-                sdf,
-                grid_sdf
-            );
-        }
-    }
-
-    #[test]
-    fn test_generate_rtree() {
-        let model = &easy_gltf::load("assets/suzanne.glb").unwrap()[0].models[0];
-        let vertices = model.vertices().iter().map(|v| v.position).collect_vec();
-        let indices = model.indices().unwrap();
-        let query_points = [
-            cgmath::Vector3::new(0.01, 0.01, 0.5),
-            cgmath::Vector3::new(1.0, 1.0, 1.0),
-            cgmath::Vector3::new(0.1, 0.2, 0.2),
-            cgmath::Vector3::new(1.1, 2.2, 5.2),
-            cgmath::Vector3::new(-0.1, 0.2, -0.2),
-        ];
-
-        let rtree_sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::Rtree,
-        );
-
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::Bvh(SignMethod::Normal),
-        );
-
-        for (idx, (rtree, sdf)) in rtree_sdf.iter().zip(sdf.iter()).enumerate() {
-            assert!(
-                (rtree - sdf).abs() < 0.01,
-                "{:?}: {} != {}",
-                query_points[idx],
-                rtree,
-                sdf
-            );
-        }
-    }
-
-    #[test]
-    fn test_generate_rtree_big() {
-        let model = &easy_gltf::load("assets/suzanne.glb").unwrap()[0].models[0];
-        let vertices = model.vertices().iter().map(|v| v.position).collect_vec();
-        let indices = model.indices().unwrap();
-
-        let bbox_min = vertices.iter().fold(
-            cgmath::Vector3::new(f32::MAX, f32::MAX, f32::MAX),
-            |acc, v| cgmath::Vector3 {
-                x: acc.x.min(v.x),
-                y: acc.y.min(v.y),
-                z: acc.z.min(v.z),
-            },
-        );
-        let bbox_max = vertices.iter().fold(
-            cgmath::Vector3::new(-f32::MAX, -f32::MAX, -f32::MAX),
-            |acc, v| cgmath::Vector3 {
-                x: acc.x.max(v.x),
-                y: acc.y.max(v.y),
-                z: acc.z.max(v.z),
-            },
-        );
-
-        let grid = Grid::from_bounding_box(&bbox_min, &bbox_max, [32, 32, 32]);
-        let mut query_points = Vec::new();
-        for x in 0..grid.get_cell_count()[0] {
-            for y in 0..grid.get_cell_count()[1] {
-                for z in 0..grid.get_cell_count()[2] {
-                    query_points.push(grid.get_cell_center(&[x, y, z]));
-                }
-            }
-        }
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::Rtree,
-        );
-        let grid_sdf = generate_grid_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &grid,
-            SignMethod::Normal,
-        );
-
-        let mut fails = 0;
-        // Test against generate_sdf
-        for (i, (sdf, grid_sdf)) in sdf.iter().zip(grid_sdf.iter()).enumerate() {
-            // Assert we have the same absolute value.
-            assert!(
-                (sdf.abs() - grid_sdf.abs()).abs() < 0.01,
-                "i: {}: {} {}",
-                i,
-                sdf,
-                grid_sdf
-            );
-
-            // Count sign issues.
-            if (sdf - grid_sdf).abs() > 0.01 {
-                fails += 1;
-            }
-        }
-
-        // We can expect sign to be different for some cells.
-        // This is because the rtree only outputs the closest triangle
-        // while bvh works with a subset of close triangles and refines the sign based on those.
-        assert!(
-            (fails as f32 / sdf.len() as f32) < 0.01,
-            "fails: {fails}/{}",
-            sdf.len()
-        );
-    }
-
-    #[test]
-    fn test_generate_rtree_bvh() {
-        let model = &easy_gltf::load("assets/suzanne.glb").unwrap()[0].models[0];
-        let vertices = model.vertices().iter().map(|v| v.position).collect_vec();
-        let indices = model.indices().unwrap();
-        let query_points = [
-            cgmath::Vector3::new(0.01, 0.01, 0.5),
-            cgmath::Vector3::new(1.0, 1.0, 1.0),
-            cgmath::Vector3::new(0.1, 0.2, 0.2),
-            cgmath::Vector3::new(1.1, 2.2, 5.2),
-            cgmath::Vector3::new(-0.1, 0.2, -0.2),
-        ];
-
-        let rtree_sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::RtreeBvh,
-        );
-
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::Bvh(SignMethod::Raycast),
-        );
-
-        for (idx, (rtree, sdf)) in rtree_sdf.iter().zip(sdf.iter()).enumerate() {
-            assert!(
-                (rtree - sdf).abs() < 0.01,
-                "{:?}: {} != {}",
-                query_points[idx],
-                rtree,
-                sdf
-            );
-        }
-    }
-
-    #[test]
-    fn test_generate_rtree_bvh_big() {
-        let model = &easy_gltf::load("assets/suzanne.glb").unwrap()[0].models[0];
-        let vertices = model.vertices().iter().map(|v| v.position).collect_vec();
-        let indices = model.indices().unwrap();
-
-        let bbox_min = vertices.iter().fold(
-            cgmath::Vector3::new(f32::MAX, f32::MAX, f32::MAX),
-            |acc, v| cgmath::Vector3 {
-                x: acc.x.min(v.x),
-                y: acc.y.min(v.y),
-                z: acc.z.min(v.z),
-            },
-        );
-        let bbox_max = vertices.iter().fold(
-            cgmath::Vector3::new(-f32::MAX, -f32::MAX, -f32::MAX),
-            |acc, v| cgmath::Vector3 {
-                x: acc.x.max(v.x),
-                y: acc.y.max(v.y),
-                z: acc.z.max(v.z),
-            },
-        );
-
-        let grid = Grid::from_bounding_box(&bbox_min, &bbox_max, [32, 32, 32]);
-        let mut query_points = Vec::new();
-        for x in 0..grid.get_cell_count()[0] {
-            for y in 0..grid.get_cell_count()[1] {
-                for z in 0..grid.get_cell_count()[2] {
-                    query_points.push(grid.get_cell_center(&[x, y, z]));
-                }
-            }
-        }
-        let sdf = generate_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &query_points,
-            AccelerationMethod::RtreeBvh,
-        );
-        let grid_sdf = generate_grid_sdf(
-            &vertices,
-            crate::Topology::TriangleList(Some(indices)),
-            &grid,
-            SignMethod::Raycast,
-        );
-
-        // Test against generate_sdf
-        for (i, (sdf, grid_sdf)) in sdf.iter().zip(grid_sdf.iter()).enumerate() {
-            assert!(
-                (sdf - grid_sdf).abs() < 0.01,
-                "i: {}: {} {}",
-                i,
-                sdf,
-                grid_sdf
-            );
-        }
     }
 }
